Kernel: Restructure FS structure for "modularizing" the kernel
Having all source files in the root directory of the kernel subtree
didn't make a whole lot of sense, but it did make things simple. This
new structure, while somewhat more complicated, should make everything
much, much cleaner. In the future, all "modules" for the kernel will
live in their own subdirectories, with header files located in the
global include directory. They can simply copy the makefiles from the
other subtrees to achieve the same compilation status as the rest of the
kernel tree.

Additionally, each makefile for any given kernel module cannot be
executed independently of the master makefile for the kernel subtree.
The master makefile sets variables that are necessary for the rest of
the tree to be consistent.