Build: Update makefiles to rebuild correctly
We only want builds to be triggered if there is an outdated .o file or
if there is a binary file that has not been rewritten after a newly
completed build. If a sub-system of the kernel has been updated and
rebuilt, then the entire kernel needs to be relinked and rewritten to
kernel.bin.